sesion03/requirements.txt
openai>=1.40.0
httpx>=0.27.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
requests>=2.32.3
numpy>=1.26.0
fastapi>=0.111.0
uvicorn>=0.30.0
huggingface-hub>=0.24.6

